backend/requirements.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
motor==2.5.1
python-dotenv==0.19.0
pydantic==1.8.2
passlib[bcrypt]==1.7.4
